### 2020-07-22 #2499 @Kevin-Mattheus-Moerman
@mdsteiner our system could not detect a license. Can you add a plain text `LICENSE` file to your repository? Adding it to the main repository folder is best (that way GitHub also easily discovers it and will render a clear license badge on your repo). The license needs to be an [OSI approved open source license](https://opensource.org/licenses).
