### 2020-07-22 #2499 @mdsteiner
thanks for the note! We have now added a `LICENSE` file to our repository.

As potential reviewers we think that maybe jacobsoj, navmayan, ihrke, or mmrabe (all from the linked list) might be suitable based on their descriptions on the domains/ topics they are comfortable to review.
