### 2020-07-26 #2521 @whedon
[REVIEW]: EFAtools: An R package with fast and flexible implementations of exploratory factor analysis tools
**Submitting author:** @mdsteiner (<a href="http://orcid.org/0000-0002-8126-0757">Markus Steiner</a>)
**Repository:** <a href="https://github.com/mdsteiner/EFAtools" target ="_blank">https://github.com/mdsteiner/EFAtools</a>
**Version:** v0.1.1
