### 2020-07-27 #2521 @chainsawriot
Thanks @mdsteiner & Co. for this wonderful package.

# Software paper

The paper is well written. But I don't know the policy of JOSS enough about citing "Manuscript in preparation". (Could @fboehm confirm this?) Obviously, the two authors are probably the only persons having that paper now. I can't 'see Grieder & Steiner, 2020', as per the in-text citation.

I have 2 suggestions:

1. Put it as a preprint and then cite it.
2. If it can't be put as a preprint (we know how these social science journals work...), maybe just say

    the ability to reproduce the R psych (Revelle, 2019) and SPSS (IBM, 2015) implementations of some analyses methods, as well as the inclusion of experimental implementations for these methods based on simulation analyses by the authors (pending publication).

# Package

## Functions

I think the documentation and the vignette are very easy to read. I can understand the package very quickly. All methods are well referenced. The `crayon`-based output is a nice touch.

The package is feature-rich and it deals with all aspects of EFA. I especially like the function `N_FACTORS` becuse it helps to solve the problem of choosing `n_factors`

The package tries to match the SPSS's offerings. I only have one suggestion (which might be too much to ask): It would be nice to have the ability to plot screeplot. Biplot might be nice too. Some reviewers would ask for it.

## Misc

* Please remove `.DS_Store` from your repo. I don't know why your `.gitignore` doesn't work. But the following should at least do the trick of finding all `.DS_Store` and ignore them.

```sh
find . -name .DS_Store -print0 | xargs -0 git rm -f --ignore-unmatch
```

Advice from `goodpractice::gp()`. You don't need to consider all of the points. But I think some are useful and less opinionated.

* Test coverage is 65% which is okay. But some important lines are not tested. e.g.
  * R/BARTLETT.R:107 (corner cases that require `cor.smooth`)
  * R/CD.R:236, 295, 303 (corner cases when `shared_load[1,1]` < 0)

* `exportPattern` is generated in NAMESPACE from [this line](https://github.com/mdsteiner/EFAtools/blob/7fc6837615a1b41d4e276b1897dbdada16f4dfc7/R/EFAtools-package.R#L4) Is it really needed? It also makes me wonder, do you need to export all helper functions? (e.g. `.numformat`)

* Adding URL / Bugreports to the `DESCRIPTION` file would be helpful for your users to find your github repo.
