### 2020-07-27 #2521 @fboehm
Regarding comments from @chainsawriot about the manuscript and the manuscript in preparation, it would be ideal if the authors can post a preprint and cite it, but, if that's not possible, then I like the other suggestion from @chainsawriot, too.
