+++ R/N_FACTORS.R
@@ -11,8 +11,8 @@
 #'  data.
 #' @param criteria character. A vector with the factor retention methods to
 #' perform. Possible inputs are: \code{"CD"}, \code{"EKC"}, \code{"HULL"},
-#' \code{"KGC"}, \code{"PARALLEL"}, and \code{"SMT"} (see details). By default,
-#' all factor retention methods are performed.
+#' \code{"KGC"}, \code{"PARALLEL"}, \code{"SCREE"}, and \code{"SMT"}
+#' (see details). By default, all factor retention methods are performed.
 #' @param suitability logical. Whether the data should be checked for suitability
 #' for factor analysis using the Bartlett's test of sphericity and the
 #' Kaiser-Guttmann criterion (see details). Default is \code{TRUE}.
@@ -59,14 +59,15 @@
 #' matrices are left to be 1. If using  \code{"EFA"}, eigenvalues are found on the
 #' correlation  matrices with the final communalities of an EFA solution as
 #' diagonal.
-#' @param eigen_type_KGC_PA character. Passed to \code{\link{KGC}} and
-#' \code{\link{PARALLEL}}. The same as eigen_type_HULL, but multiple inputs
+#' @param eigen_type_other character. Passed to \code{\link{KGC}},
+#' \code{\link{SCREE}}, and \code{\link{PARALLEL}}. The same as eigen_type_HULL,
+#' but multiple inputs
 #' are possible here. Default is to use all inputs, that is, \code{c("PCA",
 #' "SMC", "EFA"})
 #' @param n_factors numeric. Passed to \code{\link{PARALLEL}} (also within
 #' \code{\link{HULL}}) and to \code{\link{KGC}}. Number of factors to extract if
 #'  \code{"EFA"} is included in \code{eigen_type_HULL} or
-#'  \code{eigen_type_KGC_PA}. Default is 1.
+#'  \code{eigen_type_other}. Default is 1.
 #' @param n_datasets numeric. Passed to \code{\link{PARALLEL}} (also within
 #' \code{\link{HULL}}). The number of datasets to simulate. Default is 1000.
 #' @param percent numeric. Passed to \code{\link{PARALLEL}} (also within
@@ -96,6 +97,7 @@
 #' \item{Hull method (see \code{\link{HULL}})}
 #' \item{Kaiser-Guttman criterion (see \code{\link{KGC}})}
 #' \item{Parallel analysis (see \code{\link{PARALLEL}})}
+#' \item{Scree plot (see \code{\link{SCREE}})}
 #' \item{Sequential chi-square model tests, RMSEA lower bound, and AIC
 #' (see \code{\link{SMT}})}
 #' }
@@ -117,25 +119,25 @@
 #'
 #' # The same as above, but without "CD"
 #' nfac_wo_CD <- N_FACTORS(test_models$baseline$cormat, criteria = c("EKC",
-#'                         "HULL", "KGC", "PARALLEL", "SMT"), N = 500,
+#'                         "HULL", "KGC", "PARALLEL", "SCREE", "SMT"), N = 500,
 #'                         method = "ML")
 #'
 #' # Use PAF instead of ML (this will take a lot longer). For this, gof has
 #' # to be set to "CAF" for the Hull method.
 #' nfac_PAF <- N_FACTORS(test_models$baseline$cormat, criteria = c("EKC",
-#'                       "HULL", "KGC", "PARALLEL", "SMT"), N = 500,
+#'                       "HULL", "KGC", "PARALLEL", "SCREE", "SMT"), N = 500,
 #'                       gof = "CAF")
 #'
 #' # Do KGC and PARALLEL with only "PCA" type of eigenvalues
[…]
