+++ man/print.N_FACTORS.Rd
@@ -19,6 +19,6 @@ Print function for N_FACTORS objects
 # All criteria except "CD", with correlation matrix and fit method "ML"
 # (where needed)
 N_FACTORS(test_models$baseline$cormat, criteria = c("EKC", "HULL", "KGC",
-          "PARALLEL", "SMT"), N = 500, method = "ML")
+          "PARALLEL", "SCREE", "SMT"), N = 500, method = "ML")
 }
 }
