+++ NAMESPACE
@@ -5,6 +5,7 @@ S3method(plot,EKC)
 S3method(plot,HULL)
 S3method(plot,KGC)
 S3method(plot,PARALLEL)
+S3method(plot,SCREE)
 S3method(print,BARTLETT)
 S3method(print,CD)
 S3method(print,COMPARE)
@@ -17,6 +18,7 @@ S3method(print,LOADINGS)
 S3method(print,N_FACTORS)
 S3method(print,OMEGA)
 S3method(print,PARALLEL)
+S3method(print,SCREE)
 S3method(print,SL)
 S3method(print,SLLOADINGS)
 S3method(print,SMT)
@@ -35,6 +37,7 @@ export(KMO)
 export(N_FACTORS)
 export(OMEGA)
 export(PARALLEL)
+export(SCREE)
 export(SL)
 export(SMT)
 exportPattern("^[[:alpha:]]+")

[…]
