### 2020-07-28 0b9d11b
Minor update documentation N_FACTORS
