### 2020-07-28 41bb49d
add SCREE function in paper, add tests for cor.smooth cases
