### 2020-07-28 5239b24
write new vignette for replication of R psych and SPSS results, include link to new vignette in existing vignette
