+++ src/paf_iter.cpp
@@ -16,7 +16,6 @@ using namespace arma;
 //' @param abs_eig logical. Whether absolute eigenvalues should be used to compute the loadings.
 //' @param crit_type numeric. Whether maximum absolute differences (crit_type = 1), or sum of differences (crit_type = 2) should be used
 //' @param max_iter numeric. The number of iterations after which to end the procedure if no convergence has been reached by then.
-//' @export
 // [[Rcpp::export(.paf_iter)]]
 Rcpp::List paf_iter(arma::vec h2, double criterion, arma::mat R,
                     const int n_fac, bool abs_eig, int crit_type,
