+++ R/ROTATE_OBLQ.R
@@ -140,7 +140,7 @@
     if (!is.null(dim_names[[2]])) {
       dim_names[[2]] <- dim_names[[2]][ss_order]
     } else {
-      dim_names[[2]] <- paste0("F", 1:ncol(patt_mat))[ss_order]
+      dim_names[[2]] <- paste0("F", seq_len(ncol(patt_mat)))[ss_order]
     }
 
   }
@@ -162,7 +162,7 @@
       if (!is.null(dim_names[[2]])) {
         dim_names[[2]] <- dim_names[[2]][eig_order]
       } else {
-        dim_names[[2]] <- paste0("F", 1:ncol(patt_mat))[eig_order]
+        dim_names[[2]] <- paste0("F", seq_len(ncol(patt_mat)))[eig_order]
       }
 
     }

[…]
