+++ DESCRIPTION
@@ -36,6 +36,8 @@ Depends: R (>= 3.6.0)
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
+URL: https://github.com/mdsteiner/EFAtools
+BugReports: https://github.com/mdsteiner/EFAtools/issues
 RoxygenNote: 7.1.0
 Imports: 
     lavaan,
