### 2020-07-29 47d7bf6
added tests for COMPARE
