### 2020-07-29 807232b
implement possibility to do calcualte omegas based on second-order lavaan solution; update tests for OMEGA and SL; update news file accordingly
