+++ NEWS.md
@@ -4,25 +4,30 @@
 
 * Created new vignette *Replicate_SPSS_psych* to show replication of original `psych` and `SPSS` EFA solutions with `EFAtools`.
 
+## New Functions
+
+* Added function `SCREE()` that does a scree plot.
+
 ## Changes to Functions
 
 * `EFA()`: Updated the EFAtools type in PAF and Pomax.
 * `N_FACTORS()`: Added option to do a scree plot if "SCREE" is included in the `criteria` argument.
+* `SL()`: Now also works with a lavaan second-order solution as input (first- and second-order loadings taken directly from lavaan output).
+* `OMEGA()`: Now also works with a lavaan second-order solution as input. In this case, it does a Schmid-Leiman transformation based on the first- and second-order loadings first and computes omegas based on this Schmid-Leiman solution.
 
 ## Bug Fixes
 
 * `PARALLEL()`: Fixed a bug that occurred when using `decision_rule = "percentile"`
 
-## New Functions
-
-* Added function `SCREE()` that does a scree plot.
+## Minor Changes
+* `OMEGA()` and `SL()`: Added an error message if the entered term in `g_name` is invalid (i.e., it cannnot be found among the factor names of the entered lavaan solution).
 
 
 # EFAtools 0.1.1
 
 ## Minor Changes
 
-* Added an error message in PARALLEL if no solution has been found after 25 tries.
+* Added an error message in `PARALLEL()` if no solution has been found after 25 tries.
 
 ## Bug Fixes
 

[…]
