### 2020-07-29 539e1ed
minor updates tests KMO, ML, OMEGA_helpter, and PAF
