+++ tests/testthat/test-PAF.R
@@ -10,17 +10,22 @@ paf_spss <- .PAF(test_models$baseline$cormat, n_factors = 3, N = 500,
 paf_none <- .PAF(test_models$baseline$cormat, n_factors = 3, N = 500,
                  max_iter = 500, type = "none", init_comm = "unity",
                  criterion = 1e-4, criterion_type = "sums", abs_eigen = TRUE)
+paf_mac <- .PAF(test_models$baseline$cormat, n_factors = 3, N = 500,
+                 max_iter = 500, type = "none", init_comm = "mac",
+                 criterion = 1e-4, criterion_type = "sums", abs_eigen = TRUE)
 
 test_that("output class and dimensions are correct", {
   expect_is(paf_efatools$unrot_loadings, "LOADINGS")
   expect_is(paf_psych$unrot_loadings, "LOADINGS")
   expect_is(paf_spss$unrot_loadings, "LOADINGS")
   expect_is(paf_none$unrot_loadings, "LOADINGS")
+  expect_is(paf_mac$unrot_loadings, "LOADINGS")
 
   expect_output(str(paf_efatools), "List of 11")
   expect_output(str(paf_psych), "List of 11")
   expect_output(str(paf_spss), "List of 11")
   expect_output(str(paf_none), "List of 11")
+  expect_output(str(paf_mac), "List of 11")
 })
 
 test_that("original correlation matrix and eigenvalues are correct", {
@@ -28,21 +33,25 @@ test_that("original correlation matrix and eigenvalues are correct", {
   expect_equal(paf_psych$orig_R, test_models$baseline$cormat)
   expect_equal(paf_spss$orig_R, test_models$baseline$cormat)
   expect_equal(paf_none$orig_R, test_models$baseline$cormat)
+  expect_equal(paf_mac$orig_R, test_models$baseline$cormat)
 
   expect_equal(sum(paf_efatools$orig_eigen), ncol(test_models$baseline$cormat))
   expect_equal(sum(paf_psych$orig_eigen), ncol(test_models$baseline$cormat))
   expect_equal(sum(paf_spss$orig_eigen), ncol(test_models$baseline$cormat))
   expect_equal(sum(paf_none$orig_eigen), ncol(test_models$baseline$cormat))
+  expect_equal(sum(paf_mac$orig_eigen), ncol(test_models$baseline$cormat))
 
   expect_lt(sum(paf_efatools$init_eigen), ncol(test_models$baseline$cormat))
   expect_lt(sum(paf_psych$init_eigen), ncol(test_models$baseline$cormat))
   expect_lt(sum(paf_spss$init_eigen), ncol(test_models$baseline$cormat))
   expect_equal(sum(paf_none$orig_eigen), ncol(test_models$baseline$cormat))
+  expect_equal(sum(paf_mac$orig_eigen), ncol(test_models$baseline$cormat))
 
   expect_lt(sum(paf_efatools$final_eigen), ncol(test_models$baseline$cormat))
   expect_lt(sum(paf_psych$final_eigen), ncol(test_models$baseline$cormat))
   expect_lt(sum(paf_spss$final_eigen), ncol(test_models$baseline$cormat))
   expect_lt(sum(paf_none$final_eigen), ncol(test_models$baseline$cormat))
+  expect_lt(sum(paf_mac$final_eigen), ncol(test_models$baseline$cormat))
 })
 
 test_that("fit indices are returned correctly", {
@@ -50,6 +59,7 @@ test_that("fit indices are returned correctly", {
   expect_output(str(paf_psych$fit_indices), "List of 12")
   expect_output(str(paf_spss$fit_indices), "List of 12")
   expect_output(str(paf_none$fit_indices), "List of 12")
[…]
