### 2020-07-29 1cc0ca7
removed unnecessary code to create dimnames in VARIMAX, PROMAX, ROTATE_ORTH, ROTATE_OBLQ; minors in COMPARE and SMT
