+++ R/VARIMAX.R
@@ -117,11 +117,7 @@
 
     AV$rotmat <- AV$rotmat[ss_order, ss_order]
 
-    if (!is.null(dim_names[[2]])) {
-      dim_names[[2]] <- dim_names[[2]][ss_order]
-    } else {
-      dim_names[[2]] <- paste0("F", seq_len(ncol(AV$loadings)))[ss_order]
-    }
+    dim_names[[2]] <- dim_names[[2]][ss_order]
 
   } else if (order_type == "eigen") {
 
@@ -131,11 +127,7 @@
     AV$loadings <- AV$loadings[, eig_order]
     AV$rotmat <- AV$rotmat[eig_order, eig_order]
 
-    if (!is.null(dim_names[[2]])) {
-      dim_names[[2]] <- dim_names[[2]][eig_order]
-    } else {
-      dim_names[[2]] <- paste0("F", seq_len(ncol(AV$loadings)))[eig_order]
-    }
+    dim_names[[2]] <- dim_names[[2]][eig_order]
 
   }
 
