### 2020-07-29 e857644
updated tests for COMPARE, PAF, and ULS
