+++ tests/testthat/test-PAF.R
@@ -10,22 +10,40 @@ paf_spss <- .PAF(test_models$baseline$cormat, n_factors = 3, N = 500,
 paf_none <- .PAF(test_models$baseline$cormat, n_factors = 3, N = 500,
                  max_iter = 500, type = "none", init_comm = "unity",
                  criterion = 1e-4, criterion_type = "sums", abs_eigen = TRUE)
-paf_mac <- .PAF(test_models$baseline$cormat, n_factors = 3, N = 500,
+paf_mac_t <- .PAF(test_models$baseline$cormat, n_factors = 3, N = 500,
                  max_iter = 500, type = "none", init_comm = "mac",
-                 criterion = 1e-4, criterion_type = "sums", abs_eigen = TRUE)
+                 criterion = 1e-4, criterion_type = "max_individual", abs_eigen = TRUE)
+paf_mac_f <- .PAF(test_models$baseline$cormat, n_factors = 3, N = 500,
+                  max_iter = 500, type = "none", init_comm = "mac",
+                  criterion = 1e-4, criterion_type = "max_individual",
+                  abs_eigen = FALSE)
+paf_F1_t <- .PAF(test_models$baseline$cormat, n_factors = 1, N = 500,
+                max_iter = 500, type = "none", init_comm = "mac",
+                criterion = 1e-4, criterion_type = "max_individual", abs_eigen = TRUE)
+paf_F1_f <- .PAF(test_models$baseline$cormat, n_factors = 1, N = 500,
+               max_iter = 500, type = "none", init_comm = "mac",
+               criterion = 1e-4, criterion_type = "max_individual", abs_eigen = FALSE)
+
 
 test_that("output class and dimensions are correct", {
   expect_is(paf_efatools$unrot_loadings, "LOADINGS")
   expect_is(paf_psych$unrot_loadings, "LOADINGS")
   expect_is(paf_spss$unrot_loadings, "LOADINGS")
   expect_is(paf_none$unrot_loadings, "LOADINGS")
-  expect_is(paf_mac$unrot_loadings, "LOADINGS")
+  expect_is(paf_mac_t$unrot_loadings, "LOADINGS")
+  expect_is(paf_mac_f$unrot_loadings, "LOADINGS")
+  expect_is(paf_F1_t$unrot_loadings, "LOADINGS")
+  expect_is(paf_F1_f$unrot_loadings, "LOADINGS")
+
 
   expect_output(str(paf_efatools), "List of 11")
   expect_output(str(paf_psych), "List of 11")
   expect_output(str(paf_spss), "List of 11")
   expect_output(str(paf_none), "List of 11")
-  expect_output(str(paf_mac), "List of 11")
+  expect_output(str(paf_mac_t), "List of 11")
+  expect_output(str(paf_mac_f), "List of 11")
+  expect_output(str(paf_F1_t), "List of 11")
+  expect_output(str(paf_F1_f), "List of 11")
 })
 
 test_that("original correlation matrix and eigenvalues are correct", {
@@ -33,25 +51,37 @@ test_that("original correlation matrix and eigenvalues are correct", {
   expect_equal(paf_psych$orig_R, test_models$baseline$cormat)
   expect_equal(paf_spss$orig_R, test_models$baseline$cormat)
   expect_equal(paf_none$orig_R, test_models$baseline$cormat)
-  expect_equal(paf_mac$orig_R, test_models$baseline$cormat)
+  expect_equal(paf_mac_t$orig_R, test_models$baseline$cormat)
+  expect_equal(paf_mac_f$orig_R, test_models$baseline$cormat)
[…]
