+++ tests/testthat/test-COMPARE.R
@@ -1,12 +1,31 @@
-
 int <- COMPARE(1:10, 1:10)
 dec <- COMPARE(c(.1, .2), c(.1, .1))
 matr <- COMPARE(matrix(c(1,1,1,2), ncol = 2), matrix(c(1,1,1,1), ncol = 2))
+
+SPSS_PAF <- EFA(test_models$baseline$cormat, n_factors = 3, N = 500,
+                type = "SPSS", method = "PAF", rotation = "none")
+psych_PAF <- EFA(test_models$baseline$cormat, n_factors = 3, N = 500,
+                 type = "psych", method = "PAF", rotation = "none")
+load <- COMPARE(SPSS_PAF$unrot_loadings, SPSS_PAF$unrot_loadings)
+load_ro1 <- COMPARE(SPSS_PAF$unrot_loadings, SPSS_PAF$unrot_loadings,
+                    reorder = "names")
+load_ro2 <- COMPARE(SPSS_PAF$unrot_loadings, SPSS_PAF$unrot_loadings,
+                    reorder = "none")
+
+SPSS_PAF_1 <- EFA(test_models$baseline$cormat, n_factors = 1, N = 500,
+                type = "SPSS", method = "PAF", rotation = "none")
+psych_PAF_1 <- EFA(test_models$baseline$cormat, n_factors = 1, N = 500,
+                 type = "psych", method = "PAF", rotation = "none")
+load_F1 <- COMPARE(SPSS_PAF_1$unrot_loadings, psych_PAF_1$unrot_loadings)
 
 test_that("output class and dimensions are correct", {
   expect_is(int, "COMPARE")
   expect_is(dec, "COMPARE")
   expect_is(matr, "COMPARE")
+  expect_is(load, "COMPARE")
+  expect_is(load_ro1, "COMPARE")
+  expect_is(load_ro2, "COMPARE")
+  expect_is(load_F1, "COMPARE")
 
   expect_named(int, c("diff", "mean_abs_diff", "median_abs_diff", "min_abs_diff",
                       "max_abs_diff", "max_dec", "are_equal", "diff_corres",
@@ -17,6 +36,18 @@ test_that("output class and dimensions are correct", {
   expect_named(matr, c("diff", "mean_abs_diff", "median_abs_diff", "min_abs_diff",
                       "max_abs_diff", "max_dec", "are_equal", "diff_corres",
                       "diff_corres_cross", "g", "settings"))
+  expect_named(load, c("diff", "mean_abs_diff", "median_abs_diff", "min_abs_diff",
+                       "max_abs_diff", "max_dec", "are_equal", "diff_corres",
+                       "diff_corres_cross", "g", "settings"))
+  expect_named(load_ro1, c("diff", "mean_abs_diff", "median_abs_diff", "min_abs_diff",
+                       "max_abs_diff", "max_dec", "are_equal", "diff_corres",
+                       "diff_corres_cross", "g", "settings"))
+  expect_named(load_ro2, c("diff", "mean_abs_diff", "median_abs_diff", "min_abs_diff",
+                       "max_abs_diff", "max_dec", "are_equal", "diff_corres",
+                       "diff_corres_cross", "g", "settings"))
+  expect_named(load_F1, c("diff", "mean_abs_diff", "median_abs_diff", "min_abs_diff",
+                       "max_abs_diff", "max_dec", "are_equal", "diff_corres",
+                       "diff_corres_cross", "g", "settings"))
 
   expect_is(int$diff, "integer")
[…]
