### 2020-07-30 1d4016b
updated tests for HULL, PARALLEL, SMT, helper functions; added tests for remaining helper functions; some minor updates in these functions
