+++ tests/testthat/test-HULL.R
@@ -6,6 +6,11 @@ hull_cor_uls_CFI <- HULL(test_models$baseline$cormat, N = 500, method = "ULS",
 hull_cor_ml_nf <- suppressWarnings(HULL(test_models$baseline$cormat, N = 500,
                                         method = "ML", n_fac_theor = 12))
 
+hull_PCA <- HULL(test_models$baseline$cormat, N = 500, method = "ML",
+                 eigen_type = "PCA")
+hull_EFA <- HULL(test_models$baseline$cormat, N = 500, method = "ML",
+                 eigen_type = "EFA")
+
 hull_raw_paf <- suppressMessages(suppressWarnings(HULL(GRiPS_raw)))
 hull_raw_ml <- suppressMessages(suppressWarnings(HULL(GRiPS_raw, method = "ML")))
 hull_raw_uls <- suppressMessages(suppressWarnings(HULL(GRiPS_raw, method = "ULS")))
@@ -26,6 +31,10 @@ test_that("output class and dimensions are correct", {
   expect_output(str(hull_cor_uls), "List of 8")
   expect_is(hull_cor_uls_CFI, "HULL")
   expect_output(str(hull_cor_uls_CFI), "List of 8")
+  expect_is(hull_PCA, "HULL")
+  expect_output(str(hull_PCA), "List of 8")
+  expect_is(hull_EFA, "HULL")
+  expect_output(str(hull_EFA), "List of 8")
 
   expect_is(hull_raw_paf, "HULL")
   expect_output(str(hull_raw_paf), "List of 8")
@@ -35,16 +44,24 @@ test_that("output class and dimensions are correct", {
   expect_output(str(hull_raw_uls), "List of 8")
   expect_is(hull_raw_uls_CFI, "HULL")
   expect_output(str(hull_raw_uls_CFI), "List of 8")
+  expect_is(hull_PCA, "HULL")
+  expect_output(str(hull_PCA), "List of 8")
+  expect_is(hull_EFA, "HULL")
+  expect_output(str(hull_EFA), "List of 8")
 
   expect_output(str(hull_cor_paf$settings), "List of 7")
   expect_output(str(hull_cor_ml$settings), "List of 7")
   expect_output(str(hull_cor_uls$settings), "List of 7")
   expect_output(str(hull_cor_uls_CFI$settings), "List of 7")
+  expect_output(str(hull_PCA$settings), "List of 7")
+  expect_output(str(hull_EFA$settings), "List of 7")
 
   expect_output(str(hull_raw_paf$settings), "List of 7")
   expect_output(str(hull_raw_ml$settings), "List of 7")
   expect_output(str(hull_raw_uls$settings), "List of 7")
   expect_output(str(hull_raw_uls_CFI$settings), "List of 7")
+  expect_output(str(hull_PCA$settings), "List of 7")
+  expect_output(str(hull_EFA$settings), "List of 7")
 })
 
 test_that("n_fac_max is correctly specified", {
@@ -55,6 +72,10 @@ test_that("n_fac_max is correctly specified", {
   expect_lte(hull_cor_uls$n_fac_max,
              .det_max_factors(ncol(test_models$baseline$cormat)))
   expect_lte(hull_cor_uls_CFI$n_fac_max,
+             .det_max_factors(ncol(test_models$baseline$cormat)))
+  expect_lte(hull_PCA$n_fac_max,
+             .det_max_factors(ncol(test_models$baseline$cormat)))
+  expect_lte(hull_EFA$n_fac_max,
              .det_max_factors(ncol(test_models$baseline$cormat)))
 
   expect_lte(hull_raw_paf$n_fac_max,
[…]
