+++ tests/testthat/test-PARALLEL.R
@@ -2,12 +2,25 @@
 pa_cor <- PARALLEL(test_models$baseline$cormat, N = 500)
 pa_cor_pca <- PARALLEL(test_models$baseline$cormat, N = 500, eigen_type = "PCA")
 pa_raw <- PARALLEL(GRiPS_raw)
+pa_nodat <- PARALLEL(N = 20, n_vars = 5)
+pa_craw <- PARALLEL(test_models$baseline$cormat, N = 500, eigen_type = "PCA",
+                    decision_rule = "crawford")
+pa_perc <- PARALLEL(test_models$baseline$cormat, N = 500, eigen_type = "PCA",
+                    decision_rule = "percentile")
 
 test_that("output class and dimensions are correct", {
   expect_is(pa_cor, "PARALLEL")
   expect_output(str(pa_cor), "List of 7")
   expect_is(pa_raw, "PARALLEL")
   expect_output(str(pa_raw), "List of 7")
+  expect_is(pa_cor_pca, "PARALLEL")
+  expect_output(str(pa_cor_pca), "List of 7")
+  expect_is(pa_nodat, "PARALLEL")
+  expect_output(str(pa_nodat), "List of 7")
+  expect_is(pa_craw, "PARALLEL")
+  expect_output(str(pa_craw), "List of 7")
+  expect_is(pa_perc, "PARALLEL")
+  expect_output(str(pa_perc), "List of 7")
 })
 
 
@@ -50,8 +63,18 @@ test_that("found eigenvalues are correct", {
   expect_is(pa_raw$eigenvalues_SMC, "matrix")
   expect_is(pa_raw$eigenvalues_EFA, "matrix")
 
+  expect_is(pa_nodat$eigenvalues_PCA, "matrix")
+  expect_is(pa_nodat$eigenvalues_SMC, "matrix")
+  expect_is(pa_nodat$eigenvalues_EFA, "matrix")
+
   expect_is(pa_cor_pca$eigenvalues_PCA, "matrix")
   expect_equal(c(pa_cor_pca$eigenvalues_SMC, pa_cor_pca$eigenvalues_EFA), c(NA, NA))
+
+  expect_is(pa_craw$eigenvalues_PCA, "matrix")
+  expect_equal(c(pa_craw$eigenvalues_SMC, pa_craw$eigenvalues_EFA), c(NA, NA))
+
+  expect_is(pa_perc$eigenvalues_PCA, "matrix")
+  expect_equal(c(pa_perc$eigenvalues_SMC, pa_perc$eigenvalues_EFA), c(NA, NA))
 
 })
 
@@ -65,30 +88,19 @@ test_that("identified number of factors is correct", {
   expect_equal(pa_raw$n_fac_SMC, 1)
   expect_equal(pa_raw$n_fac_EFA, 3, tolerance = 2)
 
+  expect_equal(c(pa_nodat$n_fac_PCA, pa_nodat$n_fac_SMC, pa_nodat$n_fac_EFA),
+               rep(NA, 3))
+
   expect_equal(pa_cor_pca$n_fac_PCA, 3)
   expect_equal(c(pa_cor_pca$n_fac_SMC, pa_cor_pca$n_fac_EFA), c(NA, NA))
+
+  expect_equal(pa_craw$n_fac_PCA, 3)
+  expect_equal(c(pa_craw$n_fac_SMC, pa_craw$n_fac_EFA), c(NA, NA))
+
+  expect_equal(pa_perc$n_fac_PCA, 3)
+  expect_equal(c(pa_perc$n_fac_SMC, pa_perc$n_fac_EFA), c(NA, NA))
 })
 
-# Create singular correlation matrix for tests
[…]
