### 2020-07-30 1270408
added tests for case where cor.smooth is used for HULL and N_FACTORS; fixed bug in HULL if no number of factors > 0 lies on the HULL
