+++ tests/testthat/test-N_FACTORS.R
@@ -42,7 +42,7 @@ test_that("errors etc. are thrown correctly", {
   expect_warning(N_FACTORS(GRiPS_raw, N = 10), " 'N' was set and data entered. Taking N from data.\n")
   expect_error(N_FACTORS(cbind(x, y, z, z + 1, y + 1, x + 1)), " Correlation matrix is singular, no further analyses are performed\n")
   expect_warning(N_FACTORS(test_models$baseline$cormat, N = 500), " 'x' was a correlation matrix but CD needs raw data. Skipping CD.\n")
-  #expect_warning(N_FACTORS(cor_nposdef, N = 10), "Matrix was not positive definite, smoothing was done")
+  expect_warning(N_FACTORS(cor_nposdef, N = 20), "Matrix was not positive definite, smoothing was done")
 })
 
 rm(nf_grips, x, y, z, cor_nposdef)
