+++ tests/testthat/test-HULL.R
@@ -180,8 +180,7 @@ test_that("errors etc are thrown correctly", {
   expect_message(suppressWarnings(HULL(GRiPS_raw)), 'Only CAF can be used as gof if method "PAF" is used. Setting gof to "CAF"\n')
 
   expect_warning(HULL(test_models$baseline$cormat, n_fac_theor = 13, N = 500), ' Setting maximum number of factors to 12 to ensure overidentified models.\n')
-  #expect_warning(HULL(cor_nposdef, N = 20, method = "PAF"), "Matrix was not positive definite, smoothing was done")
-
+  expect_warning(HULL(cor_nposdef, N = 20, method = "ML"), "Matrix was not positive definite, smoothing was done")
 })
 
 rm(hull_cor_paf, hull_cor_ml, hull_cor_uls, hull_cor_uls_CFI, hull_raw_paf,
