### 2020-07-30 283226d
updated tests for helper and HULL; fixed bug in compare; deleted unnecessary warning message in PARALLEL
