+++ man/EFA.Rd
@@ -38,7 +38,7 @@ EFA(
   P_type = NULL,
   precision = NULL,
   order_type = NULL,
-  start_method = c("factanal", "psych"),
+  start_method = c("psych", "factanal"),
   cor_method = c("pearson", "spearman", "kendall"),
   ...
 )
@@ -135,9 +135,10 @@ rotated loadings. "ss_factors" will reorder the factors according to descending
 sum of squared factor loadings per factor. Default is \code{NULL}.}
 
 \item{start_method}{character. How to specify the starting values for the
-optimization procedure for ML. Default is "factanal" which takes the starting
-values specified in the \link[stats:factanal]{stats::factanal} function. "psych" takes the
-starting values specified in \link[psych:fa]{psych::fa}. Solutions are very similar.}
+optimization procedure for ML. Default is "psych" which takes the
+starting values specified in \link[psych:fa]{psych::fa}. "factanal" takes the
+starting values specified in the \link[stats:factanal]{stats::factanal} function.
+Solutions are very similar.}
 
 \item{cor_method}{character. Passed to \code{\link[stats:cor]{stats::cor}}.
 Default is "pearson".}
