### 2020-07-31 eddd86c
added p value for chi square, fixed bug in KMO (inverse R was not taken again after smoothing); fixed tests
