+++ paper/paper.bib
@@ -77,13 +77,6 @@
     doi = {10.1016/j.csda.2013.02.005}
   }
 
-@article{grieder2020algorithmic,
-author = {Silvia Grieder and Markus D. Steiner},
-title = {Algorithmic Jingle Jungle: A Comparison of Implementations of Principal Axis Factoring and Promax Rotation in {R} and {SPSS}},
-year = {2020},
-journal = {Manuscript in preparation}
-}
-
 @book{revelle2019psych,
   title = {psych: Procedures for Psychological, Psychometric, and Personality Research},
   author = {Revelle, W},
