+++ paper/paper.md
@@ -36,11 +36,11 @@ In the social sciences, factor analysis is a widely used tool to identify latent
 
 # Statement of Need
 
-Compared to other R packages with which EFA can be performed, *EFAtools* has several advantages, including fast implementations using *Rcpp* [@eddelbuettel2017extending; @eddelbuettel2014rcpparmadillo], more flexibility in the adjustment of implementation features, the ability to reproduce the R *psych* [@revelle2019psych] and SPSS [@ibm_spss_2015] implementations of some analyses methods, as well as the inclusion of recommended implementations for these methods based on simulation analyses [see @grieder2020algorithmic]. Finally, the package includes the implementation of the, as of yet, most comprehensive set of factor retention criteria in R, including recently developed criteria such as the Hull method [@lorenzoseva_hull_2011], comparison data [@ruscio_determining_2012], and the empirical Kaiser criterion [@braeken_empirical_2016]. As recommended by @auerswald_how_2019, multiple factor retention criteria should be examined simultaneously to check their convergence, which now is easily possible with a comprehensive function in *EFAtools* incorporating all implemented factor retention criteria for simultaneous application. Minor advantages over and above the existing implementations in R include that when intending to perform a Schmid-Leiman transformation, this can be done on an obliquely rotated solution obtained with functions from the *EFAtools* or the *psych* package instead of being forced to perform the whole EFA procedure again. Moreover, our implementation of McDonald's omegas calculations include the possibility of manual variable-to-factor correspondences (as are needed for variance partitioning for predetermined / theoretical composites) in addition to automatically determined variable-to-factor correspondences (as done, for example, in the *psych* package). Further, the *EFAtools* function to compute McDonald's omegas can easily be applied on *EFAtools* and *psych* Schmid-Leiman solutions as well as on *lavaan* [@rosseel_lavaan_2012] bifactor or single factor solutions (including solutions from multiple group analyses).
+Compared to other R packages with which EFA can be performed, *EFAtools* has several advantages, including fast implementations using *Rcpp* [@eddelbuettel2017extending; @eddelbuettel2014rcpparmadillo], more flexibility in the adjustment of implementation features, the ability to reproduce the R *psych* [@revelle2019psych] and SPSS [@ibm_spss_2015] implementations of some analyses methods (see vignette *Replicate SPSS and R psych results with EFAtools*), as well as the inclusion of recommended implementations for these methods based on simulation analyses by the authors (pending publication). Finally, the package includes the implementation of the, as of yet, most comprehensive set of factor retention criteria in R, including recently developed criteria such as the Hull method [@lorenzoseva_hull_2011], comparison data [@ruscio_determining_2012], and the empirical Kaiser criterion [@braeken_empirical_2016]. As recommended by @auerswald_how_2019, multiple factor retention criteria should be examined simultaneously to check their convergence, which now is easily possible with a comprehensive function in *EFAtools* incorporating all implemented factor retention criteria for simultaneous application. Minor advantages over and above the existing implementations in R include that when intending to perform a Schmid-Leiman transformation, this can be done on an obliquely rotated solution obtained with functions from the *EFAtools* or the *psych* package instead of being forced to perform the whole EFA procedure again. Moreover, our implementation of McDonald's omegas calculations include the possibility of manual variable-to-factor correspondences (as are needed for variance partitioning for predetermined / theoretical composites) in addition to automatically determined variable-to-factor correspondences (as done, for example, in the *psych* package). Further, the *EFAtools* function to compute McDonald's omegas can easily be applied on *EFAtools* and *psych* Schmid-Leiman solutions as well as on *lavaan* [@rosseel_lavaan_2012] second-order, bifactor, and single factor solutions (including solutions from multiple group analyses).
 
 # Development and Purpose
 
[…]
