### 2020-08-01 6a0b2fa
add FACTOR_SCORES function including tests
