+++ NAMESPACE
@@ -28,6 +28,7 @@ export(CD)
 export(COMPARE)
 export(EFA)
 export(EKC)
+export(FACTOR_SCORES)
 export(HULL)
 export(KGC)
 export(KMO)
