### 2020-08-01 8daef5a
updated factor_corres documentation and added tests
