+++ R/RcppExports.R
@@ -1,13 +1,10 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-#' Perform the iterative PAF procedure
+#' Compute number of non-matching indicator-to-factor correspondences
 #'
-#' Function called from within PAF so usually no call to this is needed by the user.
-#' Provides a C++ implementation of the PAF procedure
-#'
-#' @param x numeric matrix. The initial communality estimates.
-#' @param y numeric matrix. The convergence criterion to use.
+#' @param x numeric matrix. A matrix of pattern coefficients.
+#' @param y numeric matrix. A second matrix of coefficients.
 #' @param thresh numeric. The threshold to classify a pattern coefficient as substantial.
 .factor_corres <- function(x, y, thresh = 0.3) {
     .Call(`_EFAtools_factor_corres`, x, y, thresh)
