### 2020-08-01 b52dd59
updated COMPARE and added new tests
