### 2020-08-01 ecd466b
minor updates vignette
