+++ vignettes/Replicate_SPSS_psych.Rmd
@@ -21,7 +21,7 @@ if (!requireNamespace("psych", quietly = TRUE)) {
 
 ```
 
-This vignette demonstrates the replication of exploratory factor analysis (EFA) results (specifically, principal axis factoring [PAF] with subsequent promax rotation) from the SPSS `FACTOR` algorithm and from the `fa` function from the `psych` R package.
+This vignette demonstrates the replication of exploratory factor analysis (EFA) results (specifically, principal axis factoring [PAF] with subsequent promax rotation) from the SPSS `FACTOR` algorithm and from the `fa()` function from the `psych` R package.
 For a general introduction to the `EFAtools` package, please see the [EFAtools](EFAtools.html "EFAtools") vignette. Same as in the EFAtools vignette, we will use the DOSPERT data set for this demonstration as well (see `?DOSPERT` for details).
 
 First load the needed packages EFAtools and psych (original SPSS results for some data sets are available in the EFAtools package).
@@ -37,9 +37,11 @@ First, we will fit an EFA with PAF and without rotation using the `EFA` function
 
 ```{r}
 # EFAtools::EFA with type = "psych" without rotation
-EFA_psych_paf <- EFA(DOSPERT$cormat, n_factors = 10, N = DOSPERT$N, type = "psych")
+EFA_psych_paf <- EFA(DOSPERT$cormat, n_factors = 10, N = DOSPERT$N,
+                     type = "psych")
 # EFAtools::EFA with type = "SPSS" without rotation
-EFA_SPSS_paf <- EFA(DOSPERT$cormat, n_factors = 10, N = DOSPERT$N, type = "SPSS")
+EFA_SPSS_paf <- EFA(DOSPERT$cormat, n_factors = 10, N = DOSPERT$N,
+                    type = "SPSS")
 ```
 
 As a next step, we fit an EFA with the same configurations (PAF and no rotation) using the `fa` function from `psych` with the same data set.
@@ -59,7 +61,6 @@ COMPARE(EFA_psych_paf$unrot_loadings, psych_paf$loadings)
 # Compare loadings from SPSS and EFAtools::EFA with type = "SPSS"
 COMPARE(EFA_SPSS_paf$unrot_loadings, SPSS$DOSPERT$paf_load)
 
-COMPARE(psych_paf$loadings, SPSS$DOSPERT$paf_load)
 ```
 
 To see that this close match was not trivial, we can look at the match between the original R `psych` and `SPSS` solutions.
@@ -80,11 +81,11 @@ Now we confirmed the replication of PAF results without rotation, we can continu
 ## Fit the models
 
 # EFAtools::EFA with type = "psych" with varimax rotation
-EFA_psych_var <- EFA(DOSPERT$cormat, n_factors = 10, N = DOSPERT$N, type = "psych", 
-                     rotation = "varimax")
+EFA_psych_var <- EFA(DOSPERT$cormat, n_factors = 10, N = DOSPERT$N,
+                     type = "psych", rotation = "varimax")
 # EFAtools::EFA with type = "SPSS" with varimax rotation
-EFA_SPSS_var <- EFA(DOSPERT$cormat, n_factors = 10, N = DOSPERT$N, type = "SPSS", 
-                    rotation = "varimax")
+EFA_SPSS_var <- EFA(DOSPERT$cormat, n_factors = 10, N = DOSPERT$N,
+                    type = "SPSS", rotation = "varimax")
 # psych::fa with varimax rotation
 psych_var <- psych::fa(DOSPERT$cormat, nfactors = 10, n.obs = DOSPERT$N,
                        fm = "pa", rotate = "varimax")
[…]
