### 2020-08-01 7b2a6c9
updated NEWS.md
