### 2020-08-01 a92b5dc
updated paper, vignette, and news file
