### 2020-08-01 97faf39
minor updates vignettes
