### 2020-08-01 0b0512e
minor update paper
