+++ tests/testthat/test-RcppExports.R
@@ -7,7 +7,7 @@ y_mat <- matrix(c(.1, .1, .3,
                    0,  0,  0), ncol = 3, byrow = TRUE)
 test_that(".factor_corres works", {
   expect_equal(.factor_corres(x_mat, y_mat)$diff_corres, 1)
-  expect_equal(.factor_corres(x_mat, y_mat)$diff_corres_cross, 1)
+  expect_equal(.factor_corres(y_mat, x_mat)$diff_corres_cross, 1)
 
   expect_equal(.factor_corres(x_mat,
                               matrix(0, ncol = 3, nrow = 3))$diff_corres, 2)
