### 2020-08-01 #2521 @mdsteiner
Thank you @chainsawriot for the fast review and the many helpful comments and suggestions! Please find below our responses to your raised points.

# Software paper

>The paper is well written. But I don't know the policy of JOSS enough about citing "Manuscript in preparation". (Could @fboehm confirm this?) Obviously, the two authors are probably the only persons having that paper now. I can't 'see Grieder & Steiner, 2020', as per the in-text citation.

>I have 2 suggestions:

>    - Put it as a preprint and then cite it.
>    - If it can't be put as a preprint (we know how these social science journals work...), maybe just say
the ability to reproduce the R psych (Revelle, 2019) and SPSS (IBM, 2015) implementations of some analyses methods, as well as the inclusion of experimental implementations for these methods based on simulation analyses by the authors (pending publication).

**Response:**
As the paper is not yet in a form we are comfortable to share as a preprint, we would, for now, like to opt for the second suggestion made by the reviewer. We changed the respective sentence in the paper accordingly. However, we hope to have a shareable version ready in time to still be able to include a reference to a preprint here.

To demonstrate that our package enables replication of the R psych and SPSS solutions, we added a vignette to the package and refer to it in the paper as well.

# Package

## Functions

>I think the documentation and the vignette are very easy to read. I can understand the package very quickly. All methods are well referenced. The crayon-based output is a nice touch.

>The package is feature-rich and it deals with all aspects of EFA. I especially like the function N_FACTORS becuse it helps to solve the problem of choosing n_factors

>The package tries to match the SPSS's offerings. I only have one suggestion (which might be too much to ask): It would be nice to have the ability to plot screeplot. Biplot might be nice too. Some reviewers would ask for it.

**Response:**
Thank you for this feedback! We include a scree plot in the `KGC` and `PARALLEL` functions, but these plots are not shown if `KGC` or `PARALLEL` are called within the `N_FACTORS` function. To facilitate getting a scree plot, we now added a function `SCREE` which just creates a scree plot / scree plots of the eigenvalues determined with either "PCA", "SMC" or "EFA", or a combination of them. This function can also be called in `N_FACTORS` now, just like the other factor retention criteria.

Furthermore, we also added a function `FACTOR_SCORES` which is a wrapper for `psych::factor.scores` to be used directly with an output from `EFA`.

We also like the idea of adding the option to do a biplot and would like to consider this as a future enhancement of our package.

## Misc

>Please remove .DS_Store from your repo. I don't know why your .gitignore doesn't work. But the following should at least do the trick of finding all .DS_Store and ignore them.
`find . -name .DS_Store -print0 | xargs -0 git rm -f --ignore-unmatch`

**Response:**
Thank you for this very helpful suggestion, we now removed all .DS_Store.

>Advice from goodpractice::gp(). You don't need to consider all of the points. But I think some are useful and less opinionated.

>Test coverage is 65% which is okay. But some important lines are not tested. e.g.
R/BARTLETT.R:107 (corner cases that require cor.smooth)
R/CD.R:236, 295, 303 (corner cases when shared_load[1,1] < 0)

**Response:**
Thank you for making us aware of that. We did not know the goodpractice package and think it is really helpful. We now added new tests for some helper functions that were not tested before, and updated many of the existing tests to cover more different cases. Test coverage is now 74.7%. We do not test print and plot functions (which make up most of the untested lines) as well as some very minor cases and cases which would bear the risk of unstable results (e.g., the occurence of Heywood cases or non-convergence when random data are generated).

>exportPattern is generated in NAMESPACE from this line Is it really needed? It also makes me wonder, do you need to export all helper functions? (e.g. .numformat)

**Response:**
Thank you for pointing this out! We now deleted the exportPattern and export all functions explicitly. This way, the helper functions remain hidden.

>Adding URL / Bugreports to the DESCRIPTION file would be helpful for your users to find your github repo.

**Response:**
Thank you for this suggestion! We have now added the respective lines in the DESCRIPTION file.
