### 2020-08-03 #2521 @fboehm
Thank you, @chainsawriot ! Just to verify - are you content with the documentation around community guidelines?

Thanks again!
