### 2020-08-03 #2521 @chainsawriot
@fboehm The documentation around community guidelines looks fine to me. Really sorry that I have forgotten to check the box. The box is now checked. Thank you very much!
