### 2020-08-17 #2521 @jacobsoj
Sorry for the delay. I am one of those people juggling work and young kids w/o daycare. I should be able to finish going through the package today and submit my review by tomorrow.
Jill

Jill A. Jacobson, Ph.D. Office Phone  613-533-2847
Associate Professor, Social-Personality Program
Lab Phone 613-533-6000 x75417
Queen's University      FAX  613-533-2499
Department of Psychology        Email  jill.jacobson@queensu.ca
Kingston, ON  K7L 3N6
CANADA

http://www.queensu.ca/psychology/People/Faculty/Jill-Jacobson
