### 2020-08-17 #2521 @fboehm
Thanks, @jacobsoj ! I really appreciate your assistance with this.

Just to clarify - JOSS reviews are a little different from those at many journals. You'll want to see which items you can check off in the checklist above as you review the software and manuscript. For any boxes that you can't check right now, you'll want to tell the authors why you can't check the box, ie, what is missing or what is not working. The authors will then fix what you indicated. When they're satisfied that they've fixed the issues, they'll ask you to verify that you're satisfied with their work. The review is considered complete once all boxes are checked.

Thank you again, and please let me know how I might assist you as you work through the review.
