### 2020-08-20 4a132d2
bugfix print.COMPARE
