+++ DESCRIPTION
@@ -38,7 +38,7 @@ Encoding: UTF-8
 LazyData: true
 URL: https://github.com/mdsteiner/EFAtools
 BugReports: https://github.com/mdsteiner/EFAtools/issues
-RoxygenNote: 7.1.0
+RoxygenNote: 7.1.1
 Imports: 
     lavaan,
     psych,
