### 2020-08-20 b35b908
updated SPSS varimax implementation
