### 2020-08-20 12ce5b6
temporary setting for server
