### 2020-08-20 8a6464c
Revert "temporary setting for server"
This reverts commit 12ce5b68c062e855b14e05eadeb6d9be61840875.
