### 2020-08-20 b8d1a98
updated SPSS varimax implementation
