### 2020-08-22 57dc1ea
add argument varimax_type everywhere, change argument name kaiser to normalize, precision argument no longer type-dependent, update tests accordingly
