+++ R/EFA.R
@@ -65,7 +65,7 @@
 #' \code{NULL} (default), the value for promax depends on the specified type.
 #' For simplimax, \code{nrow(L)}, where L is the matrix of unrotated loadings,
 #' is used by default.
-#' @param kaiser logical. If \code{TRUE}, kaiser normalization is
+#' @param normalize logical. If \code{TRUE}, a kaiser normalization is
 #' performed before the specified rotation. Default is \code{TRUE}.
 #' @param P_type character. This specifies how the target
 #' matrix P is computed in promax rotation. If "unnorm" it will use the
@@ -73,11 +73,11 @@
 #' This is also used in the psych and stats packages. If "norm" it will use the
 #' normalized target matrix as used in SPSS. Default is \code{NULL}.
 #' @param precision numeric. The tolerance for stopping in the rotation
-#' procecdure. This is passed to the "eps" argument of the
-#' \code{\link[stats:varimax]{stats::varimax}} and the \code{GPArotation} functions.
-#' If left \code{NULL} (default), the precision is set according to the specified
-#' \code{type} for varimax and promax rotation or is set to 10^-5 for all other
-#' rotations.
+#' procedure. Defaul is 10^-5 for all rotation methods.
+#' @param varimax_type character. The type of the varimax rotation performed.
+#' If "svd", singular value decomposition is used, as \link[stats:varimax]{stats::varimax} does. If "kaiser", the varimax procedure performed in SPSS is used.
+#' This is the original procedure from Kaiser (1958), but with slight alterations
+#' in the varimax criterion (see Grieder & Steiner, 2020). Default is \code{NULL}.
 #' @param order_type character. How to order the factors. "eigen" will reorder
 #' the factors according to the largest to lowest eigenvalues of the matrix of
 #' rotated loadings. "ss_factors" will reorder the factors according to descending
@@ -123,29 +123,29 @@
 #' other setting combinations tested in simulation studies in Grieder & Steiner
 #' (2020), which is why this type is used as a default here.
 #'
-#' For varimax, the values of \code{precision} and \code{order_type} depend on
+#' For varimax, the values of \code{varimax_type} and \code{order_type} depend on
 #' the \code{type} argument.
 #'
 #' \code{type = "EFAtools"} will use the following argument specification:
-#' \code{precision = 1e-5, order_type = "eigen"}.
+#' \code{varimax_type = "svd", order_type = "eigen"}.
 #'
 #' \code{type = "psych"} will use the following argument specification:
-#' \code{precision = 1e-5, order_type = "eigen"}.
+#' \code{varimax_type = "svd", order_type = "eigen"}.
 #'
 #' \code{type = "SPSS"} will use the following argument specification:
-#' \code{precision = 1e-5, order_type = "ss_factors"}.
+#' \code{varimax_type = "kaiser", order_type = "ss_factors"}.
 #'
-#' For promax, the values of \code{P_type}, \code{precision},
+#' For promax, the values of \code{P_type},
 #' \code{order_type}, and \code{k} depend on the \code{type} argument.
 #'
 #' \code{type = "EFAtools"} will use the following argument specification:
-#' \code{P_type = "norm", precision = 1e-5, order_type = "eigen", k = 4}.
+#' \code{P_type = "norm", order_type = "eigen", k = 4}.
 #'
 #' \code{type = "psych"} will use the following argument specification:
-#' \code{P_type = "unnorm", precision = 1e-5, order_type = "eigen", k = 4}.
+#' \code{P_type = "unnorm", order_type = "eigen", k = 4}.
 #'
 #' \code{type = "SPSS"} will use the following argument specification:
[…]
