### 2020-08-22 9efb21d
minor fixes
