### 2020-08-27 922cc14
bugfix print.KMO
