+++ NEWS.md
@@ -21,7 +21,9 @@ methods.
     * Created an argument "varimax_type" that is set according to the specified type, but that can also be specified individually. With type R psych and EFAtools, the stats::varimax is called by default (`varimax_type = "svd"`), with type SPSS, the reproduced SPSS varimax implementation is used (`varimax_type = "kaiser"`).
     * Renamed the `kaiser` argument (controls if a Kaiser normalization is done or not) into `normalize` to avoid confusion with the `varimax_type` argument specifications.
 * `ML()`: Changed default start method to "psych".
-* `N_FACTORS()`: Added option to do a scree plot if "SCREE" is included in the `criteria` argument.
+* `N_FACTORS()`:
+    * Added option to do a scree plot if "SCREE" is included in the `criteria` argument.
+    * Added a progress bar.
 * `OMEGA()`: Now also works with a lavaan second-order solution as input. In this case, it does a Schmid-Leiman transformation based on the first- and second-order loadings first and computes omegas based on this Schmid-Leiman solution.
 * `SL()`: Now also works with a lavaan second-order solution as input (first- and second-order loadings taken directly from lavaan output).
 
