### 2020-08-27 926d4a9
Update N_FACTORS.Rd
