### 2020-08-27 1a0550a
fixed documentation N_FACTORS
