### 2020-08-30 #2521 @fboehm
@jacobsoj - I just wanted to check in to see how things are going. I recognize that this is your first JOSS review, so it might be a little confusing. We primarily use the checklist that's available above to review the software and manuscript. Please let me know if you have any questions or concerns. Thanks again!!
