### 2020-09-01 6ef763d
update paper and references
