+++ paper/paper.bib
@@ -143,9 +143,9 @@
   number = {2}
 }
 
-@Article{rosseel_lavaan_2012,
+@article{rosseel_lavaan_2012,
     title = {{lavaan}: An {R} Package for Structural Equation Modeling},
-    author = {Yves Rosseel},
+    author = {Rosseel, Yves},
     journal = {Journal of Statistical Software},
     year = {2012},
     volume = {48},
@@ -153,4 +153,12 @@
     pages = {1--36},
     url = {http://www.jstatsoft.org/v48/i02/},
     doi = {10.18637/jss.v048.i02}
+  }
+
+  @book{grieder_algorithmic_2020,
+    title = {Algorithmic jingle jungle: A comparison of implementations of {Principal} {Axis} {Factoring} and promax rotation in {R} and {SPSS}},
+    author = {Grieder, S. and Steiner, M. D.},
+    year = {2020},
+    doi = {10.31234/osf.io/7hwrm},
+    publisher = {{PsyArXiv}},
   }
