+++ paper/paper.md
@@ -9,6 +9,7 @@ tags:
 - comparison of implementations
 date: "19 July 2020"
 output:
+  pdf_document: default
   html_document:
     df_print: paged
 authors:
@@ -30,17 +31,17 @@ affiliations:
 
 # Summary
 
-In the social sciences, factor analysis is a widely used tool to identify latent constructs underlying task performance or the answers to questionnaire items. Exploratory factor analysis (EFA) is a data-driven approach to factor analysis and is used to extract a smaller number of common factors that represent or explain the common variance of a larger set of manifest variables [see @costello2005best; and @watkins2018exploratory for an overview]. Several decisions have to be made in advance when performing an EFA, including the number of factors to extract, and the extraction and rotation method to be used. After a factor solution has been found, especially for data structures in the field of intelligence research where usually high, positive factor intercorrelations occur, it is useful to subject the resulting factor solution to an orthogonalization procedure to achieve a hierarchical factor solution with one general and several specific factors. From this orthogonalized, hierarchical solution, the variance can then be partitioned to estimate the relative importance of the general versus the specific factors using omega reliability coefficients [e.g., @mcdonald1999test].
+In the social sciences, factor analysis is a widely used tool to identify latent constructs underlying task performance or the answers to questionnaire items. Exploratory factor analysis (EFA) is a data-driven approach to factor analysis and is used to extract a smaller number of common factors that represent or explain the common variance of a larger set of manifest variables [see @costello2005best and @watkins2018exploratory for an overview]. Several decisions have to be made in advance when performing an EFA, including the number of factors to extract, and the extraction and rotation method to be used. After a factor solution has been found, especially for data structures in the field of intelligence research where usually high, positive factor intercorrelations occur, it is useful to subject the resulting factor solution to an orthogonalization procedure to achieve a hierarchical factor solution with one general and several specific factors. From this orthogonalized, hierarchical solution, the variance can then be partitioned to estimate the relative importance of the general versus the specific factors using omega reliability coefficients [e.g., @mcdonald1999test].
 
 *EFAtools* is an R package [@R2018R] that enables fast and flexible analyses in an EFA framework, from tests for suitability of the data for factor analysis and factor retention criteria to hierarchical factor analysis with Schmid-Leiman transformation [@schmid1957development] and McDonald's omegas [e.g., @mcdonald1999test]. The package's core functionalities are listed in Table 1. 
 
 # Statement of Need
 
[…]
