### 2020-09-01 e61b9d9
update paper and references
