+++ paper/paper.bib
@@ -82,7 +82,7 @@
   author = {Revelle, W},
   organization = {Northwestern University},
   address = {Evanston, IL},
-  year = {2019},
+  year = {2020},
   url = {https://CRAN.R-project.org/package=psych},
   version = {1.9.12}
 }
@@ -92,7 +92,7 @@
     author = {{R Core Team}},
     organization = {R Foundation for Statistical Computing},
     address = {Vienna, Austria},
-    year = {2019},
+    year = {2020},
     url = {https://www.R-project.org/}
   }
 
