### 2020-09-01 #2521 @mdsteiner
Thanks again @chainsawriot for the helpful comments! We have now uploaded a preprint of our paper where we report the simulation analyses concerning the default implementation of PAF and promax used in the `EFA()` function and have adapted our citation in the manuscript by citing the preprint as you suggested.
