### 2020-09-01 #2521 @chainsawriot
@mdsteiner Thanks for keeping me updated. It is really nice that the preprint is available for all of us. I hope that the review of both papers (this JOSS paper and your preprint) can go faster.
