+++ man/EFA.Rd
@@ -96,7 +96,7 @@ Default is \code{NULL}.}
 \item{criterion_type}{character. Type of convergence criterion used for
 PAF. "max_individual" selects the maximum change in any of the
 communalities from one iteration to the next and tests it against the
-specified criterion. This is also used by SPSS. "sums" takes the difference of
+specified criterion. This is also used by SPSS. "sum" takes the difference of
 the sum of all communalities in one iteration and the sum of all communalities
 in the next iteration and tests this against the criterion. This procedure is
 used by the \code{\link[psych:fa]{psych::fa}} function. Default is \code{NULL}.}
@@ -208,11 +208,11 @@ For PAF, the values of \code{init_comm}, \code{criterion}, \code{criterion_type}
 and \code{abs_eigen} depend on the \code{type} argument.
 
 \code{type = "EFAtools"} will use the following argument specification:
-\code{init_comm = "smc", criterion = .001, criterion_type = "sums",
+\code{init_comm = "smc", criterion = .001, criterion_type = "sum",
 abs_eigen = TRUE}.
 
 \code{type = "psych"} will use the following argument specification:
-\code{init_comm = "smc", criterion = .001, criterion_type = "sums",
+\code{init_comm = "smc", criterion = .001, criterion_type = "sum",
 abs_eigen = FALSE}.
 
 \code{type = "SPSS"} will use the following argument specification:
@@ -303,13 +303,13 @@ EFAtools_oblim <- EFA(test_models$baseline$cormat, n_factors = 3, N = 500,
 PAF_none <- EFA(test_models$baseline$cormat, n_factors = 3, N = 500,
                 type = "none", method = "PAF", rotation = "none",
                 max_iter = 500, init_comm = "mac", criterion = 1e-4,
-                criterion_type = "sums", abs_eigen = FALSE)
+                criterion_type = "sum", abs_eigen = FALSE)
 
 # Add a promax rotation
 PAF_pro <- EFA(test_models$baseline$cormat, n_factors = 3, N = 500,
                type = "none", method = "PAF", rotation = "promax",
                max_iter = 500, init_comm = "mac", criterion = 1e-4,
-               criterion_type = "sums", abs_eigen = FALSE, k = 3,
+               criterion_type = "sum", abs_eigen = FALSE, k = 3,
                P_type = "unnorm", precision= 1e-5, order_type = "eigen",
                varimax_type = "svd")
 
