+++ README.md
@@ -60,21 +60,21 @@ library(EFAtools)
 N_FACTORS(test_models$baseline$cormat, N = 500, method = "ML")
 #> Warning in N_FACTORS(test_models$baseline$cormat, N = 500, method = "ML"): ! 'x' was a correlation matrix but CD needs raw data. Skipping CD.
 #> 
-#> ── Tests for the suitability of the data for factor a
+#> ── Tests for the suitability of the data for factor analys
 #> 
 #> Bartlett's test of sphericity
 #> 
 #> ✓ The Bartlett's test of sphericity was significant at an alpha level of .05.
 #>   These data are probably suitable for factor analysis.
 #> 
-#>   𝜒²(153) = 2173.28, p = 0
+#>   𝜒²(153) = 2173.28, p < .001
 #> 
 #> Kaiser-Meyer-Olkin criterion (KMO)
 #> 
 #> ✓ The overall KMO value for your data is marvellous with 0.916.
 #>   These data are probably suitable for factor analysis.
 #> 
-#> ── Number of factors suggested by the different facto
+#> ── Number of factors suggested by the different factor ret
 #> 
 #> ◌ Comparison data: NA
 #> ◌ Empirical Kaiser criterion: 2
@@ -86,10 +86,15 @@ N_FACTORS(test_models$baseline$cormat, N = 500, method = "ML")
 #> ◌ Kaiser-Guttman criterion with EFA: 1
 #> ◌ Parallel analysis with PCA: 3
 #> ◌ Parallel analysis with SMC: 3
-#> ◌ Parallel analysis with EFA: 6
+#> ◌ Parallel analysis with EFA: 3
 #> ◌ Sequential 𝜒² model tests: 3
 #> ◌ Lower bound of RMSEA 90% confidence interval: 2
 #> ◌ Akaike Information Criterion: 3
+```
+
+<img src="man/figures/README-unnamed-chunk-5-1.png" width="100%" /><img src="man/figures/README-unnamed-chunk-5-2.png" width="100%" />
+
+``` r
 
 # A type SPSS EFA to mimick the SPSS implementation with
 # promax rotation
@@ -101,7 +106,7 @@ EFA_SPSS
 #> 
 #> EFA performed with type = 'SPSS', method = 'PAF', and rotation = 'promax'.
 #> 
-#> ── Rotated Loadings ─────────────────────────────────
+#> ── Rotated Loadings ──────────────────────────────────────
 #> 
 #>       F1      F2      F3  
 #> V1   -.048    .035    .613
@@ -123,14 +128,14 @@ EFA_SPSS
 #> V17   .664   -.037   -.027
 #> V18   .555    .004    .050
 #> 
-#> ── Factor Intercorrelations ─────────────────────────
+#> ── Factor Intercorrelations ──────────────────────────────
 #> 
 #>       F1      F2      F3  
 #> F1    1.000   0.617   0.648
 #> F2    0.617   1.000   0.632
 #> F3    0.648   0.632   1.000
 #> 
-#> ── Variances Accounted for ──────────────────────────
+#> ── Variances Accounted for ───────────────────────────────
 #> 
 #>                       F1      F2      F3  
 #> SS loadings           4.907   0.757   0.643
@@ -139,7 +144,7 @@ EFA_SPSS
 #> Prop Comm Var         0.778   0.120   0.102
 #> Cum Prop Comm Var     0.778   0.898   1.000
 #> 
[…]
