### 2020-09-11 #2521 @jacobsoj
This package has a lot of useful features particularly the N_FACTORS and COMPARE() functions. The sequential chi-square model test on its own would save a lot of time, but then being able to evaluate across so many different extraction methods makes doing an EFA even more efficient and why I look forward to teaching my students about this package.

I found the examples in the vignette easy to follow and implement. To further test the package, I also used my own data for which I previously had run an EFA using the psych package, so I knew what to expect. All the functions worked as described with no problems, and users will appreciate the additional interpretation information provided in the output.

The suggestions proposed by the other reviewer were helpful, so I just had 2 suggestions based on the guidelines provided by the checklist:

For Documentation, I didn't see the Community guidelines in the paper or vignette (i.e., answer to: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.) On GitHub, it does say: "If you want to contribute or report bugs, please open an issue on GitHub or email us…", which might be sufficient although an email isn't provided in the text.

Overall it is well-written, but this one sentence from the Summary is quite complex and might be better broken into 2: "After a factor solution has been found, especially for data structures in the field of intelligence research where usually high, positive factor intercorrelations occur, it is useful to subject the resulting factor solution to an orthogonalization procedure to achieve a hierarchical factor solution with one general and several specific factors. " Perhaps instead: "After a factor solution has been found, it is useful to subject the resulting factor solution to an orthogonalization procedure to achieve a hierarchical factor solution with one general and several specific factors. This situation especially applies to data structures in the field of intelligence research where usually high, positive factor intercorrelations occur."
