### 2020-09-12 c0eee77
added email addresses to readme
