### 2020-09-12 #2521 @mdsteiner
Thank you @jacobsoj  for your review and your suggestions. We are happy to hear that the package was well received and that it will be used for teaching purposes!

Concerning community guidelines, we have now added our email addresses in the README, such that the sentence you referenced now reads: "If you want to contribute or report bugs, please open an issue on GitHub or email us at markus.d.steiner@gmail.com or silvia.grieder@gmail.com." Moreover, the contact information is provided in the package DESCRIPTION file (available through `packageDescription("EFAtools")` in R). There, we also include the repository URL and a BugReports URL, in line with the suggestion made by @chainsawriot.

Concerning the sentence in the summary in the paper: Thank you for this helpful suggestion! We agree that this facilitates understanding and have adopted your proposed wording in the paper.

Thanks again for your helpful comments!
