### 2020-09-14 #2521 @fboehm
@jacobsoj - I see that you've omitted checking two boxes in the checklist. With the modifications that @mdsteiner has made, are you now comfortable checking the boxes? Please do check them if you're comfortable doing so. We won't proceed until you're satisfied with the community guidelines and the quality of the writing.

Thanks again!
