### 2020-09-14 #2521 @jacobsoj
I am now comfortable checking those boxes, but it isn't letting me check them now. Is there something I need to do to allow me to edit them?
