### 2020-09-15 #2521 @fboehm
@mdsteiner - The reviewers have recommended publication of your submission. Before we can do that, I need to you to: 1. examine the article pdf proof for errors and 2. archive the package with, for example, zenodo or figshare. Please record and report the version number and archive DOI in this comments thread. Double-check that the metadata for the archived package has the same title and authors list as your JOSS manuscript.
Thanks again!
