+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: EFAtools
 Title: Fast and Flexible Implementations of Exploratory Factor Analysis Tools
-Version: 0.1.1.9000
+Version: 0.2.0
 Authors@R: 
     c(person(given = "Markus",
              family = "Steiner",
