### 2020-09-16 #2521 @mdsteiner
Thanks @fboehm! We have checked the pdf proof and everything looks fine. The package is now archived as version 0.2.0 on zenodo with the archive DOI [10.5281/zenodo.4032509](https://zenodo.org/record/4032509)

Thanks again @chainsawriot and @jacobsoj for the many helpful comments and suggestions!
